requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
mysql-connector-python==8.3.0
Werkzeug==3.0.1 
python-dotenv==1.0.1